Correcting verbs & resources flags
adding two new flags, "verbs" and "resources" to replace "verb" and "resource".
In order not to break things, old flags are markes as deprecated and will "map" to the new flags transparently for the user.

When needed, it will be possible to remove the deprecated flags and to get rid of the normalizeParamsRun function.

Unit tests have been updated to test the old deprecated flags, and the new ones.

Signed-off-by: Luc Dandoy <luc.dandoy@gmail.com>